Suppression de tout l'historique Suppression d'un élément de l'historique Fenêtre pour validation de la suppression de l'historique Ajout d'une image à la fenêtre de l'application

+++ src/main/java/cooptool/business/controllers/browsingHistoryManagement/DisplayHistoryController.java
@@ -2,24 +2,23 @@ package cooptool.business.controllers.browsingHistoryManagement;
 
 import cooptool.business.ViewLoader;
 import cooptool.business.ViewPath;
-import cooptool.business.facades.MentoringDemandFacade;
 import cooptool.business.facades.PostFacade;
 import cooptool.models.objects.MentoringDemand;
+import cooptool.models.objects.Post;
 import cooptool.models.objects.User;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.collections.FXCollections;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Button;
-import javafx.scene.control.TableCell;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.TextFieldTableCell;
+import javafx.stage.Modality;
 
 import java.net.URL;
 import java.time.format.DateTimeFormatter;
 import java.util.Locale;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 public class DisplayHistoryController implements Initializable {
@@ -27,22 +26,42 @@ public class DisplayHistoryController implements Initializable {
     @FXML
     Button deleteAllButton;
     @FXML
-    TableView<MentoringDemand> historyTableView;
+    TableView<Post> historyTableView;
     @FXML
-    TableColumn<MentoringDemand, String> titlePostCol;
+    TableColumn<Post, String> titlePostCol;
     @FXML
-    TableColumn<MentoringDemand, Void> deletePostCol;
+    TableColumn<Post, Void> deletePostCol;
 
     PostFacade postFacade = PostFacade.getInstance();
-    MentoringDemandFacade mentoringDemandFacade = MentoringDemandFacade.getInstance();
     User user;
 
     public void deleteAll(ActionEvent event) {
-        postFacade.deleteAllBrowsingHistory(user);
+
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setHeaderText(null);
+        alert.setContentText("Voulez-vous confirmer la suppression de votre historique ?");
+
+        alert.initOwner(ViewLoader.getInstance().getStage());
+
+        Optional<ButtonType> result = alert.showAndWait();
+
+        if (result.isPresent() && result.get().getButtonData().equals(ButtonBar.ButtonData.OK_DONE)) {
+            boolean res = postFacade.deleteAllBrowsingHistory(user);
+
+            if (res) {
+                historyTableView.getItems().clear();
+                historyTableView.refresh();
+            }
+        }
     }
 
-    public void deletePostFromHistory(ActionEvent event, MentoringDemand mentoringDemand) {
-        postFacade.deleteOneFromBrowsingHistory(user, mentoringDemand);
+    public void deletePostFromHistory(ActionEvent event, Post post) {
+        boolean res = postFacade.deleteOneFromBrowsingHistory(user, post);
+
+        if (res) {
+            historyTableView.getItems().remove(post);
+            historyTableView.refresh();
+        }
     }
 
     @Override
@@ -92,10 +111,21 @@ public class DisplayHistoryController implements Initializable {
 
         historyTableView.setOnMouseClicked(event -> {
             if (event.getClickCount() == 2) {
-                ViewLoader.getInstance().load(ViewPath.GET_MENTORING_DEMAND, historyTableView.getSelectionModel().getSelectedItem().getId());
+
+                Post selected = historyTableView.getSelectionModel().getSelectedItem();
+                ViewPath path;
+
+                if (selected instanceof MentoringDemand) {
+                    path = ViewPath.GET_MENTORING_DEMAND;
+                }
+                else {
+                    path = ViewPath.HOME;
+                }
+
+                ViewLoader.getInstance().load(path, selected.getId());
             }
         });
 
-        historyTableView.setItems(FXCollections.observableList(mentoringDemandFacade.getMentoringDemands()));
+        historyTableView.setItems(FXCollections.observableList(postFacade.getBrowsingHistory(user)));
     }
 }
